refactor: 달력을 공통 컴포넌트로 하여 celltype에 따라 UI 변화
