+++ common/Calendar/DateCellWithTag/styles.ts
@@ -1,70 +1,5 @@
 import styled from '@emotion/styled';
 
-export const Layout = styled.div`
-  padding: 48px 32px;
-  width: 100%;
-  height: 100%;
-`;
-
-export const Header = styled.div`
-  display: flex;
-  margin-top: 4px;
-  justify-content: space-between;
-  height: 40px;
-  margin-bottom: 12px;
-  div {
-    display: flex;
-    align-items: center;
-  }
-`;
-
-export const DateHeader = styled.div`
-  margin-right: 12px;
-  font-size: 28px;
-`;
-
-export const ArrowWrapper = styled.button`
-  display: flex;
-  align-items: center;
-  text-align: center;
-  justify-content: center;
-  width: 24px;
-  height: 24px;
-  border: 2px solid ${({ theme }) => theme.colors.neutral_400_b};
-`;
-
-export const WeekDate = styled.div`
-  display: flex;
-  width: 100%;
-  height: 32px;
-  align-items: center;
-  border-radius: 4px;
-  justify-content: space-between;
-  background-color: ${({ theme }) => theme.colors.neutral_200_b};
-  border: 2px solid ${({ theme }) => theme.colors.neutral_400_b};
-  margin-bottom: 12px;
-  div {
-    width: 100%;
-    text-align: center;
-  }
-`;
-
-interface CalendarContainerProps {
-  length: number;
-}
-
-export const CalendarContainer = styled.div<CalendarContainerProps>`
-  display: grid;
-  height: 90%;
-  grid-template-rows: ${(props) => `repeat(${props.length},1fr`};
-  border: 1px solid ${({ theme }) => theme.colors.neutral_400_b};
-`;
-
-export const WeekWrap = styled.div`
-  display: grid;
-  grid-template-columns: repeat(7, 1fr);
-`;
-
 interface DateCellProps {
   isCurrentMonth: boolean;
 }
@@ -136,4 +71,10 @@ export const Tag = styled.div<TagProps>`
     };
     return colors[color];
   }};
+
+  @media (max-width: 1632px) {
+    font-size: 12px;
+    padding: 8px;
+    gap: 2px;
+  }
 `;
