+++ pages/FineBook/index.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import MiniCalendar from './components/Calendar';
+import Calendar from '../../common/Calendar';
 import DetailFine from './components/DetailFine';
 import * as Style from './styles';
 
 const FineBook = () => {
   return (
     <Style.Layout>
-      <MiniCalendar />
+      <Calendar cellType="Mark" />
       <DetailFine />
     </Style.Layout>
   );
